Disable macOS build for now due to compilation error
The current build fails to compile due to mismatching prototypes with
`interface=AGL` (the default), but works with `interface=FREEGLUT`.

For example, `interface=AGL` requires these changes to
`include/GL/glext.h` in order to compile:

```diff
-GLAPI void APIENTRY glTexParameterIivEXT (GLenum target, GLenum pname, const GLint *params);
-GLAPI void APIENTRY glTexParameterIuivEXT (GLenum target, GLenum pname, const GLuint *params);
+GLAPI void APIENTRY glTexParameterIivEXT (GLenum target, GLenum pname, GLint *params);
+GLAPI void APIENTRY glTexParameterIuivEXT (GLenum target, GLenum pname, GLuint *params);
```

that is, the last parameter is not `const` in the headers from Apple.